requirements.txt
aiodns==3.0.0
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
brotlipy==0.7.0
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
idna==3.2
multidict==5.1.0
pycares==4.0.0
pycparser==2.20
python-i18n==0.3.9
pytz==2021.1
PyYAML==5.4.1
typing-extensions==3.10.0.0
ujson==4.0.2
yarl==1.6.3
